Automatically detect when raid is needed (#6)
* Automatically detect when raid is needed

In the past you had to specify if the instance was raid or not raid. It now looks to see how many drives are listed which have no partitions. If this is larger then 1 then we raid is created. Otherwise the disk is just formatted and mounted.

* Removing the logic for raid/non-raid

Previously you had to specify to use raid or not and a different file
was included depending on this choice, now this is automatic.