Prevent deployment ID from changing every plan
The use of timestamp() as part of the deployment ID means that,
in effect, the deployment ID changes every time a plan is created.
Instead, we use time_static, which is the recommended way to
introduce timestamps into a resouce: the timestamp is genenated only
when the resource is created in the first place.

Fixes #42.