Add time to the providers list
In my last change, I forgot to add the "time" provider to the
provider.tf list. This still works because terraform also has
an autodetect mode for providers, but the recommended way is
to use this file.

Also sorted the providers list to keep merge conflicts to a minimum
in the future.