adopt new cluster bootstrapping guidance
Starting with v22.3, Redpanda can automatically assign node IDs to empty
nodes as they join the cluster, and there is no longer a requirement to
configure a single root node to start the cluster.

This commit updates the templates and tasks to use seeds-driven
bootstrap, where all seed servers are expected to be configured with an
identical list of seed servers. Once all seed servers are up they verify
with each other that they all agree on the seeds, and collectively start
a cluster with all seed servers as the initial set of nodes in the
controller Raft group.

Admittedly, this commit doesn't do anything to ensure that the installed
version of Redpanda actually supports the new mode. In the short term,
we can address versioning with documentation and project branching.

An argument can be made that all our configs could just go directly into
the defaults.j2 template, but I opted to continue to use the `rpk
redpanda config bootstrap` command for the time being.