switched name to dodge stranded assets
Found that buildkite was canceling builds if a new commit came in which was stranding terraform assets (mostly the key and sg). I've turned that off. I've also set buildkite to skip building commits when new commits exist and the build hasnt started. Ideally I'd also have a way to limit concurrency to 1 for all builds on the pipeline but I'm not sure the concurrency: flag does that ( see : https://buildkite.com/docs/pipelines/controlling-concurrency )