add collection support to repo
Adds support for an Ansible Collection plus instructions for building it. Also fixes an issue in the readme and the license.

Not guaranteed to work on first publish, but I'm leery of publishing to galaxy without review first -- builds already work fine. After I've validated that the collection can publish and be used without issue I'll add automation around it.

I'm fairly sure that the first use will fail because node exporter isn't included, but I want to both validate that and check with Rogger on a better way to get that stuff.

add collection support to repo

Adds support for an Ansible Collection plus instructions for building it. Also fixes an issue in the readme and the license.

Not guaranteed to work on first publish, but I'm leery of publishing to galaxy without review first -- builds already work fine. After I've validated that the collection can publish and be used without issue I'll add automation around it.

I'm fairly sure that the first use will fail because node exporter isn't included, but I want to both validate that and check with Rogger on a better way to get that stuff.

Update create-ca.yml

fixing various impediments to uploading collection

remove reference to terraform