initial pass at gcp ci
Includes the GCP module and the first pass at testing around it -- still need to actually complete the buildkite bit and make sure that the GCP_CREDS passthru is legit in buildkite.

Currently the standard tiered storage testing works with ubuntu, haven't gotten a clean run with the private proxy in this format -- got an error about focal that is probably related to the ancient image we're using by default in the GCP module since I just copied that bit over from the OG stuff.