fixes and improvements
CI now works when run locally, still need to validate that it runs on buildkite. Chief issue is the passing in of GCP creds -- not sure what I have will pass muster tbh. Most likely will need to rebuild client image to accept GCP_CREDS as an env var and then use that directly as an ENV var in the script. But want to validate this first :sweat_smile:

Also did some code stuff

* Loosened some firewall rules
* bumped the basic ubuntu version to be less ancient
* fixed the build script and passed it correctly in buildkite
* allowed passing in the squid_acl so that the proxy correctly handles passthru. This need is created by the difference in subnetting techniques between GCP and AWS